api_gateway/gateway/requirements.txt
Flask==2.0.1
PyJWT==2.8.0
Werkzeug==2.2.2
requests==2.26.0
gunicorn==20.1.0
